==> scripts/01_data_preprocessing.Rmd <==
---
title: "Olink_QC"
author: "María Bueno Álvez"
output: html_document
date: "2024-03-26"
editor_options: 
  chunk_output_type: console
---

```{r, include=FALSE}
source("scripts/functions/functions_utility.R")
source("scripts/functions/functions_visualization.R")
source("scripts/functions/themes_palettes.R")

raw_data <- read_delim("data/Updated datafiles/VL-3453_ALLVOS_MEDECA_NPX_2023-08-23.csv", delim = ";")
```


# QC Warnings

We start by looking at the number of warnings per sample and assay.

```{r, warning=FALSE}
warnings_sample <-
  raw_data |>  
  filter(QC_Warning == 'WARN')  |>  
  group_by(SampleID) |> 
  summarize(num_flagged = n()) |>  
  arrange(desc(num_flagged)) |>  
  mutate(fraction_flagged = num_flagged / nrow(raw_data |> distinct(OlinkID)))|> 
  arrange(desc(num_flagged)) 

# Find samples with 50% or more warnings
high_warning_sample <-
  warnings_sample |> 
  filter(fraction_flagged > 0.5)

warnings_sample |> 
  mutate(Exclude = ifelse(SampleID %in% high_warning_sample$SampleID, "Yes", "No")) |> 
  ggplot(aes("Sample", fraction_flagged, label = SampleID, color = Exclude)) +
  geom_beeswarm() +
  geom_hline(yintercept = 0.5, linetype = 'dotted', col = 'grey')+
  geom_text(aes(label = SampleID), nudge_y = 0.01, check_overlap = T, show.legend = F) +
  scale_color_manual(values = c("No" = "grey30", "Yes" = "orangered")) +
  theme_hpa() +
  xlab("") +
  ggtitle("Fraction of warnings per sample")

ggsave(savepath("warnings_sample.png"), h = 6, w = 5)
```

There is **1 sample with > 50% QC warnings**, we remove it from the data.


```{r, warning=FALSE}
data <-
  raw_data |> 
  filter(!SampleID %in% high_warning_sample$SampleID,
         QC_Warning == "PASS")
```

After filtering, **`r length(unique(data$SampleID))` samples** remain in the data.

```{r, warning=FALSE}
warnings_assay <-
  raw_data |>  
  filter(QC_Warning == 'WARN')  |>  
  group_by(Assay) |> 
  summarize(num_flagged = n()) |>  
  arrange(desc(num_flagged)) |>  
  mutate(fraction_flagged = num_flagged / nrow(raw_data |> distinct(Assay)))|> 
  arrange(desc(num_flagged)) 

warnings_assay |> 
  ggplot(aes("Assay", num_flagged, label = Assay, color = fraction_flagged)) +
  geom_beeswarm() +
  geom_text_repel() +
  scale_color_viridis() +
  theme_hpa() +
  xlab("") +
  ggtitle("Number of warnings per assay")

ggsave(savepath("warnings_assay.png"), h = 6, w = 5)
```

All proteins (except TNF, IL6 and CXCL8 - replicate proteins) have < 5% of samples with QC warnings.

We select one measurement for of the three replicated proteins (TNF, IL6 and CXCL8) based on the mean NPX.

```{r, warning=FALSE, message=FALSE}
replicate_protein_rem <-
  data |> 
  group_by(OlinkID, Assay) |> 
  summarise(mean_NPX = mean(NPX, na.rm = T)) |> 
  group_by(Assay) |> 
  mutate(n_pro = n_distinct(OlinkID)) |> 
  filter(n_pro > 1) |> 
  group_by(Assay) |> 
  mutate(keep_pro = max(mean_NPX)) |> 
  mutate(remv_pro = ifelse(mean_NPX == keep_pro, 'no','yes')) |> 
  filter(remv_pro == 'yes') |> 
  pull(OlinkID)

# Remove all replicated assays except one each
data_qc <-
  data |> 
  filter(!OlinkID %in% replicate_protein_rem)
```

Correlation between the replicated assays:

```{r, warning=FALSE,message=FALSE}
replicate_assay <- 
  data |> 
  group_by(OlinkID, Assay) |> 
  summarise(mean_NPX = mean(NPX, na.rm = T)) |> 
  group_by(Assay) |> 
  mutate(n_pro = n_distinct(OlinkID)) |> 
  filter(n_pro > 1) |> 
  distinct(Assay) |> 
  pull(Assay)

replicate_cor_plots <- 
  lapply(replicate_assay, function(i) {
    
    OlinkIDs <-
      data |> 
      filter(Assay == i) |> 
      distinct(OlinkID) |> 
      pull(OlinkID)
    
    rep_wide<-
      data |> 
      filter(Assay == i) |> 
      distinct(OlinkID, NPX, SampleID) |>  
      pivot_wider(names_from = OlinkID, values_from = NPX)
    
    ggpairs(rep_wide, 2:5, title = i) +
      theme_hpa()
    
  })

replicate_cor_plots[[1]]
ggsave(savepath("tnf_cor.pdf"), h = 7, w = 7)
replicate_cor_plots[[2]]
ggsave(savepath("il6_cor.pdf"), h = 7, w = 7)
replicate_cor_plots[[3]]
ggsave(savepath("cxcl8_cor.pdf"), h = 7, w = 7)

data_qc |> 
  group_by(SampleID) |> 
  summarise(n = n_distinct(Assay)) |> 
  arrange(n)
```


# Explore outliers post QC

## PCA

### Per panel

```{r, warning=FALSE, message=FALSE}
panels <- 
  data_qc |> 
  pull(Panel) |> 
  unique()

pca_data_panels <- 
  map_df(panels, function(panel) {
    
    pca_dat <- 
      data_qc |> 
      filter(Panel == panel) |> 
      select(Sample = SampleID, Assay, NPX) 
    
    pca_data <- 
      do_pca(data = pca_dat,
             wide = F,
             impute = T) 
    
    pca_data$pca_res |> 
      mutate(Panel = panel)
    
  })

pca_data_panels |> 
  ggplot(aes(PC1, PC2)) +
  geom_point() +
  geom_text_repel(aes(label = Sample), show.legend = F) +
  facet_wrap(~Panel) +
  theme_hpa()

ggsave(savepath("pca_panels.png"), h = 8, w = 8)
```

We can highlight the top 3 outliers per panel:

```{r}
labels_panels <- 
  pca_data_panels |> 
  mutate(sum_PC1_PC2 = abs(PC1) + abs(PC2)) |> 
  group_by(Panel) |> 
  top_n(3, sum_PC1_PC2)

pca_data_panels |> 
  ggplot(aes(PC1, PC2)) +
  geom_point() +
  geom_text_repel(data = labels_panels, aes(label = SampleID), size = 3, show.legend = F) +
  facet_wrap(~Panel) +
  theme_hpa()
ggsave(savepath("pca_panels.png"), h = 8, w = 8)
```


### All proteins

```{r, warning=FALSE, message=FALSE}
pca_dat <- 
  data_qc |> 
  select(Sample = SampleID, Assay, NPX) 

pca_data <- 
  do_pca(data = pca_dat,
         wide = F,
         impute = T) 

labels <- 
  pca_data$pca_res |> 
  mutate(sum_PC1_PC2 = abs(PC1) + abs(PC2)) |> 
  top_n(3, sum_PC1_PC2)

pca_scores_panel <-
  pca_data$pca_res |> 
  ggplot(aes(PC1, PC2)) +
  geom_point() +
  geom_text_repel(data = labels, aes(label = Sample), show.legend = F) +
  theme_hpa()

# Loadings
pca_loadings <- 
  pca_data$loadings |> 
  select(Assay, Value, PC) |> 
  pivot_wider(names_from = PC,
              values_from = Value) |>
  left_join(raw_data |> 
              distinct(Assay, Panel), by = "Assay") |> 
  ggplot(aes(PC1, PC2, color = Panel)) +
  geom_point() +
  geom_text_repel(aes(label = Assay), show.legend = F) +
  scale_color_manual(values = pal_panel) +
  theme_hpa()

pca_scores_panel + pca_loadings + ggtitle("PCA biplot")
ggsave(savepath("pca_biplot.png"), h = 5, w = 9)
```


## UMAP

### Per panel

```{r, warning=FALSE, message=FALSE}
umap_data_panels <- 
  map_df(panels, function(panel) {
    
    umap_dat <- 
      data_qc |> 
      filter(Panel == panel) |> 
      select(Sample = SampleID, Assay, NPX) 

    umap_data <-  
      do_umap(data = umap_dat,
              wide = F,
              impute = T) |> 
      mutate(Panel = panel)
    
  })

umap_data_panels |> 
  ggplot(aes(UMAP1, UMAP2)) +
  geom_point() +
  facet_wrap(~Panel) +
  theme_hpa()
```

### All proteins

```{r, warning=FALSE, message=FALSE}
umap_dat <- 
  data_qc |> 
  select(Sample = SampleID, Assay, NPX) 
  
umap_data <-  
  do_umap(data = umap_dat,
          wide = F,
          impute = T)

umap_data |> 
  ggplot(aes(UMAP1, UMAP2)) +
  geom_point(alpha = 0.7, size = 2) +
  theme_hpa()
```


## NPX distribution plot

NPX distribution per panel:

```{r, warning=FALSE}
data_qc|> 
  ggplot(aes(NPX, fill = Panel)) +
  geom_density(alpha = 0.7, color = "black") +
  facet_wrap(~Panel, scales = "free") +
  scale_fill_manual(values = pal_panel) +
  theme_hpa() +
  ggtitle("NPX distirbution")

ggsave(savepath("npx_distribution.png"), h = 6, w = 7)
```

## LOD 

Samples with > 25% os proteins below LOD:

```{r, warning=FALSE, message=FALSE}
# LOD per sample
sample_lod <- 
  data_qc |> 
  mutate(under_LOD = ifelse(NPX < LOD, "Yes", "No")) |> 
  group_by(SampleID, under_LOD) |> 
  summarise(n = n_distinct(Assay)) 

sample_order <- 
  sample_lod |> 
  filter(under_LOD == "Yes") |> 
  arrange(-n) |> 
  head(20) |> 
  pull(SampleID)

sample_lod |> 
  filter(SampleID %in% sample_order) |> 
  mutate(SampleID = factor(SampleID, levels = sample_order)) |> 
  ggplot()
```

Proteins with > 50% of samples below LOD:

```{r, warning=FALSE, message=FALSE}
# LOD per protein
lod_protein <- 
  data_qc |> 
  mutate(under_LOD = ifelse(NPX < LOD, 1, 0)) |> 
  filter(under_LOD == 1) |> 
  group_by(Assay) |> 
  summarise(n = n_distinct(SampleID)) |> 
  arrange(-n) |> 
  mutate(perc = (n/length(unique(data_qc$SampleID)))*100) 

lod_protein_more50 <- 
  lod_protein |> 
  filter(perc > 50) |>  # 107 proteins with > 50% of samples below LOD
  rename(n_under_LOD = n, perc_under_LOD = perc)

#head(lod_protein_more50, 10)
n_prots_lod <- nrow(lod_protein_more50)

lod_protein |> 
  ggplot(aes("Protein", perc)) +
  geom_quasirandom() +
  theme_hpa() +
  geom_text_repel(aes(label = Assay), size = 3) +
  xlab("")

lod_protein_more50 |> 
  ggplot(aes(fct_reorder(Assay, -perc_under_LOD), perc_under_LOD, color = perc_under_LOD, fill = perc_under_LOD)) +
  geom_bar(stat = "identity") +
  geom_text(aes(label = n_under_LOD, y = perc_under_LOD + 2.5), vjust = -0.5, size = 3, angle = 90) +
  theme_hpa(angled = T) +
  xlab("Assay")

ggsave(savepath("LOD_protein.png"), h = 7, w = 25)
```

# Remove controls

```{r}

control_samples <- 
  data_qc |> 
  filter(Sample_Type == "CONTROL") |> 
  distinct(SampleID) |> 
  pull()

final_data <- 
  data_qc |> 
  filter(!SampleID %in% control_samples) |> 
  select(SampleID, Assay, NPX)

final_data_wide <- 
  data_qc |> 
  filter(!SampleID %in% control_samples) |> 
  select(SampleID, Assay, NPX) |> 
  pivot_wider(names_from = Assay, values_from = NPX)

```

# Save final data 

```{r}
write_csv(final_data, savepath_data("olink_data", "olink_data_qc.csv"))
write_csv(final_data_wide, savepath_data("olink_data", "olink_data_wide_qc.csv"))

# Panel info
data_qc |> 
  distinct(Assay, Panel) |> 
  write_csv(savepath_data("olink_metadata", "olink_panel_info.csv"))
```
