==> scripts/02_metadata_preprocessing.Rmd <==
---
title: "Metadata preprocessing"
output: html_document
date: "2024-03-26"
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
# Read functions & packages
source("scripts/functions/functions_utility.R")
source("scripts/functions/functions_visualization.R")
source("scripts/functions/themes_palettes.R")

# Read data
data <- 
  import_df("data/processed/olink_data/olink_data_qc.csv") |> 
  rename(Sample = SampleID) |> 
  mutate(Sample = as.character(Sample))

metadata <- import_df("data/Exports from R after processing updated data files/all_data.xlsx")

metadata <- 
  metadata |> 
  select(!starts_with("Olink")) |>
  rename(Sample = Patient_number) |> 
  filter(Sample %in% data$Sample) |> 
  mutate(ID = as.numeric(Sample), 
         Cohort = ifelse(ID > 1000, "ALLVOS", "MEDECA")) |> 
  select(-ID)

metadata_medeca <- import_df("data/Pat 1-728 Jan 7 2025.xlsx")

metadata_allvos <- import_df("data/Cohortdata proteomics 240429 till fredrika.xlsx")
metadata_allvos_controls <- import_df("data/Proteomics Diagnoses Feb 25 to Stockholm.xlsx")
#a <- import_df("data/ALLVOS/Diagnosis Proteomics 238.xlsx")
# 
# metadata_allvos_controls |> 
#   count(Classification) 
```

# Metadata pre-processing

## MEDECA

```{r}
keep_columns <- c("Patient number", "Sex 1 male", "Age", "CUP", "Ongoing cancer known prior to DC", 
                  "Loss to follow up","Other disease than cancer", "No disease", "Autoimmune1",
                  "Inflammatory", "Infectious", "thromboembolism", "Other disease", "Metastases", 
                  "Carcinom", "Germinal cell tumors", "SCC", "Urothelial", "Neuroendocrine tumors",
                  "Adenocarcinoma", "HCC", "Kidney", "Adrenal", "Prostate", "Lung", "Breast", "Thyroid",
                  "Gynecologic","Ovarian or tubar", "Colorectal", "Urine bladder", "Merkel cell carcinoma",
                  "Small cell parotid", "Pancreas or gall bladder or bile duct", "Ventricle", 
                  "Other adenocarcinoma", "Other carcinoma such as mesothelioma", "Sarkom", "Lymfom",
                  "Melanom", "Myelom", "MDS MPN", "Other cancer", "Mortality")

meta_medeca <- 
  metadata_medeca |>
  select(keep_columns) |> 
  rename(Sample = `Patient number`) |> 
  mutate(`Loss to follow up` = ifelse(is.na(`Loss to follow up`), 0, `Loss to follow up`)) |>
  filter(Sample %in% data$Sample,
         `Ongoing cancer known prior to DC` == 0,
         `Loss to follow up` != 1) |> 
  mutate(Sex = ifelse(`Sex 1 male` == 1, "Male", "Female")) |> 
  select(-`Ongoing cancer known prior to DC`, -`Loss to follow up`, - `Sex 1 male`) |> 
  relocate(Sex, .after = "Age") |> 
  mutate(#`No disease` = ifelse(Sample == 494, 0, `No disease`), #Correction
         #`Other disease than cancer` = ifelse(Sample == "138", 1, `Other disease than cancer`), # Correction
         #`No disease` = ifelse(Sample == "138", 0, `No disease`), # Correction
          Category = case_when(`Other disease than cancer` == 0 & `No disease` == 0 ~ "Cancer",
                              `Other disease than cancer` == 0 & `No disease` == 1 ~ "No diagnosis",
                              `Other disease than cancer` == 1 & `No disease` == 0 ~ "Other diseases",
                              T ~ NA),
         Disease_type = case_when(Autoimmune1 == 1 ~ "Autoimmune",
                                  Inflammatory == 1 ~ "Inflammatory", 
                                  Infectious == 1 ~ "Infectious", 
                                  thromboembolism == 1 ~ "Other", 
                                  `Other disease` == 1 ~ "Other",
                                  Category == "Other diseases" ~ "Other", #Sample 297 & 372 changed
                                  Category == "Cancer" ~ "Cancer",
                                  Category == "No diagnosis" ~ "No diagnosis"),
         Cancer = ifelse(Disease_type == "Cancer", "Yes", "No")) |>  
  select(-Autoimmune1, -Inflammatory, -Infectious, -thromboembolism) |> 
  relocate(Category, Disease_type, .after = Sex) |> 
  mutate(Sample = as.character(Sample),
         Disease_type = factor(Disease_type, 
                               levels = c("Cancer", "No diagnosis", "Autoimmune", "Infectious", "Inflammatory", "Other")),
         Category = factor(Category, 
                           levels = c("Cancer", "No diagnosis", "Other diseases")))


meta_medeca |> 
  filter(Sample %in% data$Sample) |> 
  count(Category, Disease_type)  |> 
  ggplot(aes(Disease_type, n, fill = Category)) +
  geom_col() +
  geom_text(aes(label = n), vjust = -0.5) +
  scale_fill_manual(values = pal_category) +
  theme_hpa(angled = T) +
  ggtitle("Sample size in processed metadata")


#ggsave(savepath("metadata_medeca_category.png"), width = 6, height = 6)
```


## ALLVOS 

```{r}
meta_allvos <- 
  metadata_allvos |>
  mutate(Sample = as.character(StudyID),
         `Lost to follow up` = ifelse(is.na(`Lost to follow up`), 0, `Lost to follow up`)) |> 
  filter(`Ongoing cancer` != 1,
         `Lost to follow up` != 1,
         StudyID %in% data$Sample) |> 
  select(Sample, Age, Sex) |> 
  left_join(metadata_allvos_controls |>
              mutate(`Sample ID` = as.character(`Sample ID`)) |> 
              select(Sample = `Sample ID`, Disease_type = Classification), by = "Sample") |> 
  mutate(Cancer = ifelse(Disease_type == "Cancer", "Yes", "No"),
         Category = ifelse(Disease_type %in% c("Autoimmune", "Infectious", "Inflammatory", "Other"), "Other diseases", Disease_type),
         Disease_type = factor(Disease_type, 
                               levels = c("Cancer", "No diagnosis", "Autoimmune", "Infectious", "Inflammatory", "Other")),
         Category = factor(Category, 
                           levels = c("Cancer", "No diagnosis", "Other diseases")))



meta_allvos |> 
  filter(Sample %in% data$Sample) |> 
  count(Category, Disease_type)  |> 
  ggplot(aes(Disease_type, n, fill = Category)) +
  geom_col() +
  geom_text(aes(label = n), vjust = -0.5) +
  scale_fill_manual(values = pal_category) +
  theme_hpa(angled = T) +
  ggtitle("Sample size in processed metadata")
```

## Combine

```{r}
meta_combined <- 
  meta_medeca |> 
  select(Sample, Age, Sex, Cancer, Category, Disease_type) |> 
  mutate(Cohort = "MEDECA") |> 
  bind_rows(meta_allvos |> 
              select(Sample, Age, Sex, Cancer, Category, Disease_type) |> 
              mutate(Cohort = "ALLVOS")) 
```

## Filter data to include only samples with metadata

```{r}
data_filtered <- 
  data |> 
  filter(Sample %in% meta_combined$Sample) 

length(unique(data$Sample))
nrow(meta_allvos) + nrow(meta_medeca) 
nrow(meta_combined)
length(unique(data_filtered$Sample))
```

# Save processed data & metadata 

```{r}
dir.create("data/processed/final_data/")

write_csv(data_filtered, "data/processed/final_data/final_olink_data.csv")  
write_csv(meta_combined, "data/processed/final_data/combined_metadata.csv")  
write_csv(meta_medeca, "data/processed/final_data/medeca_metadata.csv")  
write_csv(meta_allvos, "data/processed/final_data/allvos_metadata.csv")  
```

# Cohort overview

```{r}
meta_combined  |>
  group_by(Cancer) |>
  count(Cohort) |>
  mutate(Cancer = case_when(Cancer == "Yes" ~ "Cancer",
                            Cancer == "No" ~ "No cancer")) |> 
mutate(Cancer = factor(Cancer, levels = c("Cancer", "No cancer")),
         Cohort = factor(Cohort, levels = rev(c("MEDECA", "ALLVOS")))) |>

  ggplot(aes(Cancer, n, fill = Cancer)) +
  geom_col(width = 0.5) +
  scale_fill_manual(values = c(pal_cancer[[1]], pal_cancer[[2]])) + 
  geom_text(aes(label = paste0("n = ", n), y =n + 40), vjust = 0.5, fontface = "bold", color = "grey20") +
  theme_hpa() +
  facet_wrap(~Cohort, scales = "free_y") +
  coord_flip() +
  theme(axis.title.y = element_blank(),
        axis.ticks.y = element_blank(),
        axis.line.y = element_blank())

#ggsave(savepath("n_medeca_allvos.pdf"), h = 3, w = 8)
```
