==> scripts/03_differential_expression.Rmd <==
---
title: "Differential expression analyses"
output: html_document
date: "2024-03-26"
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
# Read functions & packages
source("scripts/functions/functions_utility.R")
source("scripts/functions/functions_analyses.R")
source("scripts/functions/functions_visualization.R")
source("scripts/functions/themes_palettes.R")

# Read data
data <- import_df("data/processed/final_data/final_olink_data.csv")  
metadata <- import_df("data/processed/final_data/combined_metadata.csv")  
meta_medeca <- import_df("data/processed/final_data/medeca_metadata.csv")  
meta_allvos <- import_df("data/processed/final_data/allvos_metadata.csv")  
hpa <- import_df("data/hpa/proteinatlas.tsv")
pancancer_markers <- import_df("data/41467_2023_39765_MOESM7_ESM.xlsx")
```


# MEDECA

```{r}
# Prepare data
limma_meta_medeca <- 
  meta_medeca |> 
  select(Sample, Cancer, Age, Sex) 
  
limma_data_medeca <- 
  data |> 
  filter(Sample %in% limma_meta_medeca$Sample) |> 
  pivot_wider(names_from = Assay, 
              values_from = NPX) 

# Run differential expression analyses
medeca_de <-
  do_limma(data_wide = limma_data_medeca, 
           metadata = limma_meta_medeca,
           variable = "Cancer", 
           case = "Yes",
           correct = T) 

#saveRDS(medeca_de, savepath_data("DE", "de_medeca.rds"))
#volcano_medeca <- plot_volcano(medeca_de) + ggtitle("MEDECA")
#medeca_de <- readRDS(savepath_data("DE", "de_medeca.rds"))
```


# ALLVOS

```{r}
# Prepare data
limma_meta_allvos <- 
  meta_allvos |>
  select(Sample, Cancer, Age, Sex) 
  
limma_data_allvos <- 
  data |> 
  filter(Sample %in% limma_meta_allvos$Sample) |> 
  pivot_wider(names_from = Assay, 
              values_from = NPX) 

 # Run differential expression analyses
allvos_de <-
  do_limma(data_wide = limma_data_allvos, 
           metadata = limma_meta_allvos,
           variable = "Cancer", 
           case = "Yes",
           correct = T) 
 
#saveRDS(allvos_de, savepath_data("DE", "de_allvos.rds"))
#volcano_allvos <- plot_volcano(allvos_de) + ggtitle("ALLVOS")
#allvos_de <- readRDS(savepath_data("DE", "de_allvos.rds"))
```


# Direct comparison

```{r}
allvos_de |> 
  select(Assay, logFC_ALLVOS = logFC, pval_ALLVOS = adj.P.Val) |> 
  left_join(medeca_de |> 
              select(Assay, logFC_MEDECA = logFC, pval_MEDECA = adj.P.Val), by = "Assay") |> 
  ggplot(aes(logFC_ALLVOS, logFC_MEDECA)) +
  geom_point() +
  geom_text_repel(aes(label = Assay), size = 2) +
  geom_vline(xintercept = 0, color = "grey") +
  geom_hline(yintercept = 0, color = "grey") +
  theme_hpa()

#ggsave(savepath("proteins_MEDECA_ALLVOS.pdf"), h = 8, w = 8)
```



# Volcano plot

```{r}
medeca_de |> 
  left_join(allvos_de |> 
              select(Assay, sig_ALLVOS = sig)) |> 
  mutate(Significant = case_when(sig == "significant up" & sig_ALLVOS == "significant up" ~ "Upregulated in both",
                                 sig == "significant down" & sig_ALLVOS == "significant down" ~ "Downregulated in both",
                                   TRUE ~ "Not significant or discordant")) |> 
    ggplot(aes(x = logFC, y = -log10(adj.P.Val), color = Significant, label = Assay)) +
    geom_point(size = 1, alpha = 0.4, show.legend = F) + 
    geom_text_repel( size = 2, show.legend = F) +
    geom_hline(yintercept = -log10(0.05), linetype = 'dashed', color = "darkgrey") +
    geom_vline(xintercept = -0.5, linetype = 'dashed', color = "darkgrey") +
    geom_vline(xintercept = 0.5, linetype = 'dashed', color = "darkgrey") +
    scale_color_manual(values = unname(pal_de)) +
    theme_hpa() +   
    theme(axis.text = element_text(size = 8),
          legend.position = "top") 

ggsave(savepath("MEDECA_ALLVOS_volcano.pdf"), h = 4, w = 4)
```

# Top proteins

```{r}
# Select top replicated proteins
replicated_proteins <- 
  medeca_de |> 
  left_join(allvos_de |> 
              select(Assay, 
                     sig_ALLVOS = sig, 
                     FC_ALLVOS = logFC,
                     pval_ALLVOS = adj.P.Val)) |> 
  mutate(Significant = case_when(sig == "significant up" & sig_ALLVOS == "significant up" ~ "Upregulated in both",
                                 sig == "significant down" & sig_ALLVOS == "significant down" ~ "Downregulated in both",
                                   TRUE ~ "Not significant or discordant")) |> 
  filter(Significant == "Upregulated in both") |> 
  select(Assay, 
         FC_MEDECA = logFC, 
         FC_ALLVOS,
         pval_MEDECA = adj.P.Val,
         pval_ALLVOS) 

top_proteins <- 
  replicated_proteins |> 
  head(7) |> 
  pull(Assay)
           
plot_beeswarm_cohorts(data = data,
                      metadata = metadata,
                      variable = "Cancer",
                      palette = pal_cancer,
                      proteins = top_proteins)

#ggsave(savepath("top_proteins.pdf"), h = 12, w = 5)
```

# Logistic regression

```{r}
proteins <- 
  data |> 
  distinct(Assay) |> 
  pull()
```

## MEDECA

```{r}
lr_data_medeca <- 
  data |> 
  filter(Sample %in% limma_meta_medeca$Sample) |> 
  left_join(meta_medeca |> 
              select(Sample, Age, Sex, Cancer)) |> 
  mutate(Sex = as.factor(Sex))

res_lr_medeca <- 
  map_df(proteins, function(protein) {
    do_lr_protein(data = lr_data_medeca,
                  protein = protein)
  }) |> 
  mutate(p.value.adjusted = p.adjust(p.value, method = "BH")) 

res_lr_medeca_or <- 
  map_df(proteins, function(protein) {
    do_lr_protein_or(data = lr_data_medeca,
                  protein = protein)
  }) |> 
  mutate(p.value.adjusted = p.adjust(p.value, method = "BH")) 
```


## ALLVOS

```{r}
lr_data_allvos <- 
  data |> 
  filter(Sample %in% limma_meta_allvos$Sample)|> 
  left_join(meta_allvos |> 
              select(Sample, Age, Sex, Cancer)) |> 
   mutate(Sex = as.factor(Sex))

res_lr_allvos <- 
  map_df(proteins, function(protein) {
    do_lr_protein(data = lr_data_allvos,
                  protein = protein)
  }) |> 
  mutate(p.value.adjusted = p.adjust(p.value, method = "BH"))  

res_lr_allvos_or <- 
  map_df(proteins, function(protein) {
    do_lr_protein_or(data = lr_data_allvos,
                  protein = protein)
  }) |> 
  mutate(p.value.adjusted = p.adjust(p.value, method = "BH"))  
```

## Forest plot

```{r}
data_forest <- 
  res_lr_medeca |> 
  filter(term == "NPX") |> 
  mutate(Cohort = "MEDECA") |> 
  bind_rows(res_lr_allvos |> 
              filter(term == "NPX") |> 
              mutate(Cohort = "ALLVOS")) |> 
  filter(Assay %in% replicated_proteins$Assay) |> 
  mutate(Assay= factor(Assay, levels = (replicated_proteins$Assay)),
         Cohort = factor(Cohort, levels = rev(c("MEDECA", "ALLVOS")))) |> 
  arrange(Cohort, Assay)

ggforestplot::forestplot(
  df = data_forest,
  name = Assay,
  estimate = estimate,
  se = std.error,
  pvalue = p.value,
  psignif = 0.05,
  colour = Cohort,
  xlab = "log(OR)",
  logodds = TRUE
) +
  scale_color_manual(values = c("grey70", "grey40")) +
  theme(legend.position = "top")

ggsave(savepath("forestplot_lr.pdf"), h = 8, w = 4)
```

## Prepare file

```{r}
res_lr_medeca_or |>
  filter(term == "NPX") |>
  mutate(Cohort = "MEDECA") |>
  bind_rows(res_lr_allvos_or |>
              filter(term == "NPX") |>
              mutate(Cohort = "ALLVOS")) |>
  filter(Assay %in% replicated_proteins$Assay) |>
  mutate(Assay = factor(Assay, levels = (replicated_proteins$Assay)),
         Cohort = factor(Cohort, levels = (c("MEDECA", "ALLVOS")))) |>
  arrange(Cohort, Assay) |>
  select(
    Cohort,
    Assay,
    OR = estimate,
    std.error,
    statistic,
    p.value,
    p.value.adjusted,
    conf.low,
    conf.high
  ) |>
  mutate(
    OR = round(OR, 2),
    std.error = round(std.error, 2),
    statistic = round(statistic, 2),
    p.value = sprintf("%.2e", p.value),
    p.value.adjusted = sprintf("%.2e", p.value.adjusted),
    conf.low = round(conf.low, 2),
    conf.high = round(conf.high, 2)
  ) |>
  write_csv(savepath("lr_results.csv"))
```


# Functional characterization

```{r}
proteins_func <- 
  medeca_de |> 
  select(Assay, logFC, adj.P.Val) |> 
  mutate(Cohort = "MEDECA") |> 
  bind_rows(allvos_de |> 
              select(Assay, logFC, adj.P.Val) |> 
              mutate(Cohort = "ALLVOS")) |> 
  filter(Assay %in% replicated_proteins$Assay)

secreted <-
  hpa |>
  distinct(Secretome.location) |>
  filter(grepl("Secreted", Secretome.location)) |>
  pull()

de_prots <- 
  proteins_func |> 
  mutate(Assay= factor(Assay, levels = rev(replicated_proteins$Assay)),
         Cohort = factor(Cohort, levels = c("MEDECA", "ALLVOS"))) |>
  ggplot(aes(Cohort, Assay, color = logFC, size = -log10(adj.P.Val))) +
  geom_point() +
  scale_color_paletteer_c(palette = "ggthemes::Red") +
  theme_hpa(angled = T) +
  theme(axis.ticks.y = element_blank(),
        legend.position = "left") +
  ylab("") +
  xlab("")

#de_prots <- 
  res_lr_medeca_or |> 
  filter(term == "NPX") |> 
  mutate(Cohort = "MEDECA") |> 
  bind_rows(res_lr_allvos_or |> 
              filter(term == "NPX") |> 
              mutate(Cohort = "ALLVOS")) |> 
  mutate(Assay= factor(Assay, levels = rev(replicated_proteins$Assay)),
         Cohort = factor(Cohort, levels = c("MEDECA", "ALLVOS"))) |>
    filter(!is.na(Assay)) |> 
  ggplot(aes(x = estimate, y = Assay, color = Cohort)) +
  geom_point(size = 2.5,                        position = position_dodge(width = 0.3)) +
  geom_errorbarh(aes(xmin = conf.low, xmax = conf.high), height = 0.2,                        position = position_dodge(width = 0.3)) +
  geom_vline(xintercept = 1, linetype = "dashed", color = "grey") +
  scale_x_log10() +  # log scale for ORs
  scale_color_manual(values = c("grey40", "grey70")) +
  labs(x = "Odds Ratio (log scale)") +
  theme_hpa()



 #  ggplot(aes(estimate, Assay, color = Cohort)) +
 #  geom_point() +
 #  geom_segment(aes(x = conf.low, xend = conf.high)) +
 # # scale_color_paletteer_c(palette = "ggthemes::Red") +
 #  theme_hpa(angled = T) +
 #  theme(axis.ticks.y = element_blank(),
 #        legend.position = "left") +
 #  ylab("") +
 #  xlab("")


characterization_prots <- 
  hpa |> 
  filter(Gene %in% replicated_proteins$Assay) |> 
  select(Gene,
         RNA.tissue.specificity, 
         Secretome.location,
         Secretome.function,
         Biological.process,
         Molecular.function,
         Disease.involvement) |> 
  mutate(Pan_cancer_marker = ifelse(Gene %in% pancancer_markers$Protein, "Yes", "No"),
         Secreted = ifelse(Secretome.location %in% secreted, "Yes", "No"),
         Tissue_enriched = ifelse(RNA.tissue.specificity == "Tissue enriched", "Yes", "No"),
         Cancer_related = case_when(grepl("Cancer",Disease.involvement) ~ "Yes",
                                    grepl("Proto-oncogene",Disease.involvement) ~ "Yes",
                                    T ~ "No"),
         Gene = factor(Gene, levels = rev(replicated_proteins$Assay))) |> 
  select(Gene, Secreted, Tissue_enriched, Cancer_related, Pan_cancer_marker) |> 
  pivot_longer(cols = -Gene, names_to = "Type", values_to = "Value") |> 
  ggplot(aes(Type, Gene, fill = Value)) +
  geom_tile(color = "white") +
  coord_fixed() +
  scale_fill_manual(values = c("grey90", "grey30")) +
  xlab("") +
  theme_hpa(angled = T, axis_y = F)

de_prots + characterization_prots +
  plot_layout(guides = "collect")

ggsave(savepath("proteins_functional_characterization.pdf"), h = 12, w = 6)
```

# PCA

## MEDECA

```{r}
pca_data <- 
  data |> 
  filter(Sample %in% meta_medeca$Sample,
         Assay %in% proteins_func$Assay) |> 
  pivot_wider(names_from = Assay, values_from = NPX)

medeca_pca<- do_pca(data = pca_data, 
                    wide = T)

medeca_variance <- 
  medeca_pca$pca_variance |> 
  filter(component %in% c(1,2))

medeca_variance_pc1 <- medeca_variance %>% filter(component == 1) %>% pull(value)
medeca_variance_pc2 <- medeca_variance %>% filter(component == 2) %>% pull(value)


pca_medeca <- 
  medeca_pca$pca_res |> 
  left_join(meta_medeca, by = "Sample") |> 
  ggplot(aes(PC1, PC2, color = Category)) +
  geom_point() +
  stat_ellipse(geom = "polygon", alpha = 0.2, aes(fill = Category), color = NA) +
  scale_color_manual(values = pal_category) +
  scale_fill_manual(values = pal_category) +
  theme_hpa() +
  labs(
    x = paste0("PC1 (", round(medeca_variance_pc1, 2), "% variance explained)"),
    y = paste0("PC2 (", round(medeca_variance_pc2, 2), "% variance explained)")
  )

final_pca_medeca <- 
  ggMarginal(
  pca_medeca,
  type = "density",  # Marginal plot type (density plot)
  margins = "x",  # Apply to both PC1 and PC2
  size = 5,          # Adjust the size of the marginal plots
  groupColour = TRUE, # Color by Category
  groupFill = TRUE    # Fill by Category
) |> 
  as.ggplot()
```

## ALLVOS

```{r}
pca_data_allvos <- 
  data |> 
  filter(Sample %in% meta_allvos$Sample,
         Assay %in% proteins_func$Assay) |> 
  pivot_wider(names_from = Assay, values_from = NPX)

allvos_pca<- do_pca(data = pca_data_allvos, 
                      wide = T)

allvos_variance <- 
  allvos_pca$pca_variance |> 
  filter(component %in% c(1,2))

allvos_variance_pc1 <- allvos_variance %>% filter(component == 1) %>% pull(value)
allvos_variance_pc2 <- allvos_variance %>% filter(component == 2) %>% pull(value)

pca_allvos <- 
  allvos_pca$pca_res |> 
  left_join(meta_allvos, by = "Sample") |> 
  ggplot(aes(PC1, PC2, color = Category)) +
  geom_point() +
  stat_ellipse(geom = "polygon", alpha = 0.2, aes(fill = Category), color = NA) +
  scale_color_manual(values = pal_category) +
  scale_fill_manual(values = pal_category) +
  #  stat_density_2d(aes(fill = ..level..), geom = "polygon", alpha = 0.3) + 
  theme_hpa()+
  labs(
    x = paste0("PC1 (", round(allvos_variance_pc1, 2), "% variance explained)"),
    y = paste0("PC2 (", round(allvos_variance_pc2, 2), "% variance explained)")
  )

final_pca_allvos <- 
  ggMarginal(
  pca_allvos,
  type = "density",  # Marginal plot type (density plot)
  margins = "x",  # Apply to both PC1 and PC2
  size = 5,          # Adjust the size of the marginal plots
  groupColour = TRUE, # Color by Category
  groupFill = TRUE    # Fill by Category
) |> 
  as.ggplot()
```

## Combine


```{r}
final_pca_medeca/final_pca_allvos
ggsave(savepath("pca_allvos_medeca.pdf"), width = 6, height = 10)
```

## Explore PC1 - MEDECA & ALLVOS

```{r}
medeca_pca$loadings |> 
  filter(PC == "PC1") |> 
  mutate(dataset = "MEDECA") |> 
  bind_rows(allvos_pca$loadings |> 
              filter(PC == "PC1")  |> 
              mutate(dataset = "ALLVOS")) |> 
  mutate(dataset = factor(dataset, levels = c("MEDECA", "ALLVOS"))) |>
  ggplot(aes(reorder_within(Assay, Value, dataset), Value, fill = Value)) +
  geom_col() +
  coord_flip() +
  facet_wrap(~dataset, ncol = 1, scales = "free") +
  theme_hpa() +
  xlab("") +
  scale_fill_distiller(type = "div") +
  scale_x_reordered() +
  theme(axis.ticks.y = element_blank(),
        axis.line.y = element_blank()) 

ggsave(savepath("loadings_pc1.pdf"), h = 10, w = 4)


# Compare directly
medeca_pca$loadings |> 
  filter(PC == "PC1") |> 
  select(Assay, MEDECA_value = Value) |> 
  left_join(allvos_pca$loadings |> 
              filter(PC == "PC1")  |> 
              select(Assay, ALLVOS_value = Value), by = "Assay") |> 
  ggplot(aes(MEDECA_value, ALLVOS_value)) +
  geom_point() +
  geom_text_repel(aes(label = Assay)) +
  theme_hpa()

ggsave(savepath("comparison_pc1.pdf"), h = 8, w = 8)

# Overlap
pc1_medeca <- 
  medeca_pca$loadings |> 
  filter(PC == "PC1") |> 
  pull(Assay)

pc1_allvos <- 
  allvos_pca$loadings |> 
  filter(PC == "PC1") |> 
  pull(Assay)

y <- list("MEDECA" = pc1_medeca, 
          "ALLVOS" =pc1_allvos)

plot(euler(y, shape = "ellipse"), quantities = TRUE,  alpha = 0.5) |> as.ggplot() +
  theme(plot.title = element_text(hjust = 0.5,
                                  face = "bold"))
```

## Boxplots

```{r}
proteins <- c("DPY30", "MAD1L1", "METAP1D", "FKBP4", "GRPEL1", "ZBTB17")

variable <- "Category"
palette <- pal_category

data |> 
    filter(Assay %in% proteins,
           Sample %in% metadata$Sample) |> 
    left_join(metadata |> 
                select(Sample, !!sym(variable), Cohort), by = "Sample") |> 
    mutate(Cohort = factor(Cohort, levels = c("MEDECA", "ALLVOS")),
           Category = factor(Category, levels = c("No diagnosis", "Other diseases", "Cancer")),
           Assay = factor(Assay, levels = proteins),
           Cancer = case_when(
               Category == "No diagnosis" ~ "No_0",
               Category == "Other diseases" ~ "No_1",
               Category == "Cancer" ~ "Yes"
           )) |> 
    ggplot(aes(Cancer, NPX, fill = Category, color = Category)) +
    geom_quasirandom(size = 0.5) +
    geom_boxplot(alpha = 0.5, outlier.color = NA) +
    stat_summary(fun = "median",
                 geom = "crossbar", 
                 width = 0.2,
                 colour = "grey20",
                 show.legend = F) +
    facet_grid(Assay~Cohort, scales = "free") +
    scale_color_manual(values = palette) +
    scale_fill_manual(values = palette) +
    theme_hpa() +
    theme(legend.position = "top")


ggsave(savepath("proteins_continuous.pdf"), h = 11, w = 5)
```
