==> scripts/04_cancer_classification.Rmd <==
---
title: "Cancer classification"
output: html_document
date: "2024-03-26"
editor_options: 
  chunk_output_type: console
---


```{r setup, include=FALSE}
# Read functions & packages
source("scripts/functions/functions_utility.R")
source("scripts/functions/functions_analyses.R")
source("scripts/functions/functions_visualization.R")
source("scripts/functions/themes_palettes.R")

# Read data
data <- import_df("data/processed/final_data/final_olink_data.csv")  
metadata <- import_df("data/processed/final_data/combined_metadata.csv")  
meta_medeca <- import_df("data/processed/final_data/medeca_metadata.csv")  
meta_allvos <- import_df("data/processed/final_data/allvos_metadata.csv")  
medeca_de <- readRDS(savepath_data("DE", "de_medeca.rds"))
```



# MEDECA

## Prepare data

```{r}
# Prepare ML data
ml_data <- 
  data |> 
  filter(Sample %in% meta_medeca$Sample) |> 
  select(Sample, Assay, NPX) |> 
  pivot_wider(names_from = Assay,
              values_from = NPX) |> 
  left_join(meta_medeca |> 
              select(Sample, Cancer), by = "Sample") 
```


#  Prep data

```{r}
# Split data 
set.seed(213)
medeca_split <-
  ml_data |> 
  initial_split(prop = 0.7, strata = Cancer)

medeca_train <- training(medeca_split)
medeca_test <- testing(medeca_split)
```


# Univariate filtering

```{r}
data_select_proteins <- 
  data |> 
  filter(Sample %in% medeca_train$Sample) |> 
  left_join(metadata |> 
              select(Sample, Cancer), by = "Sample")

# AUC filter
selected_proteins_auc <- 
  step_filter_auc(data = data_select_proteins,
                  cutoff = 0.7)

# NPF filter
selected_proteins_npx <- 
  step_filter_t_test(data = data_select_proteins,
                     output = "NPX",
                     cutoff = 1)

# p-value filter
selected_protein_pval <- 
  step_filter_t_test(data = data_select_proteins,
                     output = "pval",
                     cutoff = 1E-6)

# Combine
y <- list("AUC > 0.7" = selected_proteins_auc$final_list$Assay, 
          "NPX diff > 1" = selected_proteins_npx$final_list$Assay,
          "adj pval < 1E-6" = selected_protein_pval$final_list$Assay)

plot(euler(y, shape = "circle"), quantities = TRUE, fills = list(fill = c("#449395", "#EA7C68", "#FDA638")),  alpha = 0.4) |> as.ggplot() 
ggsave(savepath("selected_proteins.pdf"), width = 3, height = 3)

selected_protein_list <- union(union(y$`AUC > 0.7`, y$`NPX diff > 1`), y$`adj pval < 1E-6`)
```



# LASSO

## Subset data

```{r}
medeca_train_filtered <- 
  medeca_train |> 
  select(Sample, Cancer, selected_protein_list)

medeca_test_filtered <- 
  medeca_test |> 
  select(Sample, Cancer, selected_protein_list)
```

## Explore protein correlations

```{r}
# Look at correlations between selected proteins
cor_selected_proteins <- 
  medeca_train_filtered |> 
  select(-Cancer) |> 
  column_to_rownames("Sample") |> 
  cor(method = "spearman", use = "complete.obs")

cor_selected_proteins |> 
  as.data.frame() |> 
  rownames_to_column("Protein_1") |> 
  as_tibble() |> 
  pivot_longer(names_to = "Protein_2",
               values_to = "cor",
               cols = c(2:18)) |> 
  filter(Protein_1 != Protein_2) |> 
  arrange(-cor)

cor_selected_proteins |> 
  pheatmap() |> 
  as.ggplot()

ggsave(savepath("corr_selected_proteins.png"), h = 6, w = 6)
```

## Run lasso


```{r}
# Run lasso pipeline
ml_medeca_filter<- 
  do_glmnet(variable = "Cancer",
           case = "Yes", 
           split_train = medeca_train_filtered, 
           split_test = medeca_test_filtered) 

#saveRDS(ml_medeca_filter, savepath_data("ML", "ml_medeca_filter.rds"))
```

## Performance

```{r}
# Explore performance
ml_medeca_filter$performance
ml_medeca_filter$roc_curve |> autoplot() + theme_hpa()
```

## Protein importance

```{r}
importance <- 
  ml_medeca_filter$important_proteins |> 
  select(-Sign) |> 
  bind_rows(tibble("Variable" = "Lower_limit",
                   "Importance" = 0)) |> 
  mutate(Importance = scales::rescale(Importance, c(0,1))) |> 
  filter(Importance > 0,
         Variable != "Lower_limit") 

importance |> 
  ggplot(aes(fct_reorder(Variable, Importance), Importance)) +
  geom_segment(aes(x = fct_reorder(Variable, Importance), 
                   xend = fct_reorder(Variable, Importance), 
                   y = 0, 
                   yend = Importance),
               color = "#C4C4C4",) +
  geom_point(color = "#C4C4C4") + 
  coord_flip() +
  theme_hpa(angled = T) +
  xlab("") +
  ylab("Model estimate") +
  #scale_color_manual(values = c(pal_de[[3]], pal_de[[2]])) +
  theme(axis.text.y = element_text(size = 9),
        axis.ticks.y = element_blank(),
        legend.position = "top")

ggsave(savepath("top_ml_markers.pdf"), h = 7, w = 2.5)
#ggsave(savepath("top_ml_markers.png"), h = 7, w = 2.5)
```


## Robustness selection analyses 

```{r}
# Add ALLVOS validation in the function
final_models <- map(1:50, ~ run_ml_seed(.x))

#saveRDS(final_models, savepath_data("final_models", "final_models_medeca.rds"))
#final_models <- readRDS(savepath_data("final_models", "final_models_medeca.rds"))

# Look first at selected features??
selected_proteins_seeds <- 
  map_df(1:50, function(seed) {
    
    tibble(Proteins = final_models[[seed]]$selected_proteins,
           Seed = seed)
  })

 
plot_robustness <- 
  selected_proteins_seeds |> 
  count(Proteins) |> 
  filter(n > 5) |> #nrow()
  mutate(In_model = ifelse(Proteins %in% selected_protein_list, "Yes", "No")) |> 
  ggplot(aes(x = n, y = fct_reorder(Proteins, n), fill = In_model)) + 
  geom_col() +
  geom_text(aes(label = n), hjust = -0.1) +
  scale_fill_manual(values = pal_binary) +
  theme_hpa() +
  ylab("") +
  xlab("Number of proteins") +
  ggtitle("61 proteins selected in more than 5 training sets")

# dat <- 
#   selected_proteins_seeds |> 
#   count(Seed) |>
#   mutate(Seed = as.factor(Seed))
# 
# mean_n <- 
#   dat |> 
#   pull(n) |> 
#   mean()
# 
# plot_n <- 
#   dat |> 
#   ggplot(aes(n, Seed)) +
#   geom_col() +
#   geom_text(aes(label = n), hjust = -0.1) +
#   geom_vline(xintercept = mean_n, linetype = "dashed", color = "red") +
#   theme_hpa() +
#   xlab("Number of seeds") +
#   ggtitle("Number of features pre-selected in each training set")
# 
# 
# plot_n + plot_robustness
# ggsave(savepath("robustness_analyses.png"), width = 10, height = 10)
```


## Top proteins - boxplot
 
```{r}
plot_ml <- 
  ml_medeca_filter$important_proteins |> 
  head(5)

plot_beeswarm_cohorts(proteins = plot_ml$Variable,
                      data = data |> filter(Sample %in% meta_medeca$Sample),
                      metadata = metadata,
                      variable = "Cancer",
                      palette = pal_cancer)

#ggsave(savepath("top_proteins_ml.pdf"), h = 10, w = 4)
```


# Apply to ALLVOS

```{r}
# Prepare ALLVOS data
ml_data_allvos <- 
  data |> 
  filter(Sample %in% meta_allvos$Sample) |> 
  select(Sample, Assay, NPX) |> 
  pivot_wider(names_from = Assay,
              values_from = NPX) |> 
  left_join(meta_allvos |> 
              select(Sample, Cancer), by = "Sample") |> 
  mutate(Class = case_when(Cancer == "Yes" ~ "Case",
                           Cancer!= "Yes" ~ "Control")) |> 
  mutate(Class = factor(Class, levels = c("Control", "Case"))) |> 
  select(Sample, selected_protein_list, Class) |> 
  mutate(Sample = as.character(Sample)) 

# Finalize the workflow with the best parameters
final_workflow_fitted <- 
  ml_medeca_filter$final_workflow %>%
  fit(data = ml_medeca_filter$current_training)

# Preprocess the independent data with the trained recipe
prepared_allvos_data <- 
  final_workflow_fitted %>%
  extract_recipe() %>%
  bake(new_data = ml_data_allvos)

# Generate predictions on the preprocessed data
allvos_predictions <- predict(final_workflow_fitted, 
                              new_data = prepared_allvos_data, 
                              type = "prob")

# Add predicted class and true class for evaluation
allvos_predictions <- 
  allvos_predictions %>%
  mutate(Pred = ifelse(.pred_Case > 0.5, "Case", "Control"),
         True_Class = ml_data_allvos$Class)


allvos_predictions |>
  roc_curve(truth = True_Class, .pred_Case, event_level = "second") |> 
  autoplot()

allvos_predictions |>
  roc_auc(truth = True_Class, .pred_Case, event_level = "second") 
```



# Combined

## ROC 

```{r}
auc_allvos <- 
  allvos_predictions |>
  roc_auc(truth = True_Class, .pred_Case, event_level = "second") 

auc_medeca <- 
  ml_medeca_filter$performance 

# Generate combined ROC curve
tiles <-
  tibble(specificity = c(0.2,0.2),
         sensitivity = c(0.3,0.1),
         Cohort = c("Discovery", "Replication"),
         AUC = c(round(auc_medeca$.estimate,2),
                 round(auc_allvos$.estimate,2)))


ml_medeca_filter$roc_curve |> 
  mutate(Cohort = "Discovery") |> 
  bind_rows(allvos_predictions |> 
              roc_curve(truth = True_Class, .pred_Case, event_level = "second") |> 
              mutate(Cohort = "Replication")) |> 
  ggplot(aes(
    x = 1 - specificity,
    y = sensitivity,
    group = Cohort
  )) +
  geom_path(aes(color = Cohort), size = 1) +
  geom_segment(aes(
    x = 0,
    y = 0,
    xend = 1,
    yend = 1
  ),
  colour = 'grey',
  linetype = 'dotdash') +
  geom_tile(
    width = 0.2, 
    height = 0.2,
    data = tiles,
    aes(fill = Cohort),
    #alpha = 0.7,
    show.legend = F
  ) +
  geom_text(
    data = tiles,
    aes(label = AUC),
    size = 3,
    color = "white",
    show.legend = F
  ) +
  geom_text(
    label = "AUC",
    x = 0.8,
    y = 0.5,
    size = 3,
    inherit.aes = F
  ) +  # Fixed the warning
  scale_color_manual(values = c("grey40", "grey70")) +
  scale_fill_manual(values = c("grey40", "grey70")) +
  scale_x_continuous(breaks = c(0, 1)) +
  scale_y_continuous(breaks = c(0, 1)) +
  theme_hpa() +
  coord_fixed() +
  theme(legend.position = "top")

ggsave(savepath("ROC_MEDECA_ALLVOS.pdf"), h = 4, w = 5)  
```

## Probabilities

```{r}
ml_medeca_filter$predictions |> 
  ggplot(aes(.pred_Case, fill = Class,  color = Class)) +
  geom_density(alpha = 0.7, show.legend = F) +
  geom_vline(data = ml_medeca_filter$predictions  %>% group_by(Class) %>% 
               summarise(mean_pred = mean(.pred_Case, na.rm = TRUE)), 
             aes(xintercept = mean_pred, color = Class), 
             linetype = "dashed", size = 1, show.legend = F) +
  scale_fill_manual(values = c("Case" = "#E16C54", "Control" = "#A0CBC7")) + 
  scale_color_manual(values = c("Case" = "#E16C54", "Control" = "#A0CBC7")) + 
  theme_hpa() +
  xlab("Probability cancer") +
  
  allvos_predictions|> 
  ggplot(aes(.pred_Case, fill = True_Class,  color = True_Class)) +
  geom_density(alpha = 0.7) +
  geom_vline(data = allvos_predictions %>% group_by(True_Class) %>% 
               summarise(mean_pred = mean(.pred_Case, na.rm = TRUE)), 
             aes(xintercept = mean_pred, color = True_Class), 
             linetype = "dashed", size = 1, show.legend = F) +
  scale_fill_manual(values = c("Case" = "#E16C54", "Control" = "#A0CBC7")) +   
  scale_color_manual(values = c("Case" = "#E16C54", "Control" = "#A0CBC7")) + 
  theme_hpa() +
  theme(legend.position = "top") +
  xlab("Probability cancer") 

ggsave(savepath("prob_meodels.pdf"), h = 5, w = 7)

ml_medeca_filter$predictions |> 
  ggplot(aes(.pred_Case, fill = Class,  color = Class)) +
  geom_density(alpha = 0.7, show.legend = F) +
  geom_vline(data = ml_medeca_filter$predictions  %>% group_by(Class) %>% 
               summarise(mean_pred = mean(.pred_Case, na.rm = TRUE)), 
             aes(xintercept = mean_pred, color = Class), 
             linetype = "dashed", size = 1, show.legend = F) +
  scale_fill_manual(values = c("Case" = "#E16C54", "Control" = "#A0CBC7")) + 
  scale_color_manual(values = c("Case" = "#E16C54", "Control" = "#A0CBC7")) + 
  theme_hpa() +
  xlab("Probability cancer") +
  
  allvos_predictions|> 
  ggplot(aes(.pred_Case, fill = True_Class,  color = True_Class)) +
  geom_density(alpha = 0.7) +
  geom_vline(data = allvos_predictions %>% group_by(True_Class) %>% 
               summarise(mean_pred = mean(.pred_Case, na.rm = TRUE)), 
             aes(xintercept = mean_pred, color = True_Class), 
             linetype = "dashed", size = 1, show.legend = F) +
  scale_fill_manual(values = c("Case" = "#E16C54", "Control" = "#A0CBC7")) +   
  scale_color_manual(values = c("Case" = "#E16C54", "Control" = "#A0CBC7")) + 
  theme_hpa() +
  theme(legend.position = "top") +
  xlab("Probability cancer") 

#ggsave(savepath("prob_meodels.pdf"), h = 5, w = 7)

ml_medeca_filter$predictions  |> 
  select(Probability = .pred_Case,
         Class) |> 
  mutate(Class = case_when(Class == "Case" ~ "Yes",
                           Class == "Control" ~ "No")) |> 
  ggplot(aes(Class, Probability, fill = Class, color = Class)) +
  geom_quasirandom(show.legend = F) +
  geom_violin(alpha = 0.3, show.legend = F) +
  stat_summary(fun = "mean",
               geom = "crossbar", 
               width = 0.3,
               colour = "grey20",
               show.legend = F) +
   scale_fill_manual(values = c("Yes" = "#E16C54", "No" = "#A0CBC7")) +   
  scale_color_manual(values = c("Yes" = "#E16C54", "No" = "#A0CBC7")) + 
  xlab("Cancer") + 
  theme_hpa()  +

allvos_predictions |> 
  select(Probability = .pred_Case,
         True_Class) |> 
  mutate(True_Class = case_when(True_Class == "Case" ~ "Yes",
                                True_Class == "Control" ~ "No")) |> 
  ggplot(aes(True_Class, Probability, fill = True_Class, color = True_Class)) +
  geom_quasirandom(show.legend = F) +
  geom_violin(alpha = 0.3, show.legend = F) +
  stat_summary(fun = "mean",
               geom = "crossbar", 
               width = 0.3,
               colour = "grey20",
               show.legend = F) +
  scale_fill_manual(values = c("Yes" = "#E16C54", "No" = "#A0CBC7")) +   
  scale_color_manual(values = c("Yes" = "#E16C54", "No" = "#A0CBC7")) + 
  xlab("Cancer") +
  theme_hpa() 

ggsave(savepath("prob_models.pdf"), h = 4, w = 6)


```


# Stratification control groups

## Number of samples plot

```{r}
plot_medeca_train <- 
  medeca_train |> 
  select(Sample) |> 
  left_join(meta_medeca |> 
              select(Sample, Disease_type), 
            by = "Sample") |>
  count(Disease_type) |>
  filter(!Disease_type %in% c("Cancer")) |> 
  mutate(Disease_type = factor(Disease_type, levels = c("Other","No diagnosis", "Infectious", "Autoimmune", "Inflammatory", "Cancer"))) |> 
  ggplot(aes(Disease_type, n, fill = Disease_type)) +
  geom_bar(stat = "identity", show.legend = F) +
  geom_text(aes(label = n), vjust = -0.5) +
  scale_fill_manual(values = pal_controls) +
  theme_hpa(angled = T, axis_x = F) +
  ggtitle("Training set")

plot_medeca_test <- 
  medeca_test |> 
  select(Sample) |> 
  left_join(meta_medeca |> 
              select(Sample, Disease_type), 
            by = "Sample") |>
  count(Disease_type) |>
  filter(!Disease_type %in% c("Cancer")) |> 
  mutate(Disease_type = factor(Disease_type, levels = c("Other","No diagnosis", "Infectious", "Autoimmune", "Inflammatory", "Cancer"))) |> 
  ggplot(aes(Disease_type, n, fill = Disease_type)) +
  geom_bar(stat = "identity", show.legend = F) +
  geom_text(aes(label = n), vjust = -0.5) +
  scale_fill_manual(values = pal_controls) +
  theme_hpa(angled = T, axis_x = F) +
  ggtitle("Test set")
  
  
empty_plot <- ggplot() + theme_void()

plot_allvos <- 
  meta_allvos|> 
  count(Disease_type) |> 
  filter(!Disease_type %in% c("Cancer")) |> 
  mutate(Disease_type = factor(Disease_type, levels = c("Other","No diagnosis", "Infectious", "Autoimmune", "Inflammatory", "Cancer"))) |> 
  ggplot(aes(Disease_type, n, fill = Disease_type)) +
  geom_bar(stat = "identity", show.legend = F) +
  geom_text(aes(label = n), vjust = -0.5) +
  scale_fill_manual(values = pal_controls) +
  theme_hpa(angled = T, axis_x = F) 

plot_medeca_train / plot_medeca_test
ggsave(savepath("control_groups_medeca.pdf"), h = 5.5, w = 2.5)
empty_plot / plot_allvos
ggsave(savepath("control_groups_allvos.pdf"), h = 5, w = 3)

```

## ROC 

```{r}
plot_roc_tiles(predictions = ml_medeca_filter$predictions, 
               test_data = medeca_test,
               metadata = meta_medeca) 
  
ggsave(savepath("roc_curve_controls_medeca.pdf"), height = 5, width = 5)

plot_roc_tiles(predictions = allvos_predictions |> 
                 rename(Class = True_Class) |> 
                 mutate(Class = factor(Class, levels = c("Control", "Case"))), 
               test_data = ml_data_allvos,
               metadata = meta_allvos |> mutate(Sample = as.character(Sample))) 

ggsave(savepath("roc_curve_controls_allvos.pdf"), height = 5, width = 5)
```

## Densities

```{r}
plot_densities(predictions =  ml_medeca_filter$predictions,
               metadata = meta_medeca, 
               test_data = medeca_test_filtered)

ggsave(savepath("roc_controls_medeca.pdf"), height = 10, width = 3)


plot_densities(predictions =  allvos_predictions,
               metadata = meta_allvos,
               test_data = ml_data_allvos |> mutate(Sample = as.numeric(Sample))) 

ggsave(savepath("roc_controls_allvos.pdf"), height = 10, width = 3)
```


## Plot proteins across control groups

```{r}
proteins <- c("AGR2", "KRT19", "CEACAM5", "RRM2", "PARP1")


data |> 
  filter(Assay %in% proteins,
         Sample %in% metadata$Sample) |> 
  left_join(metadata |> 
              select(Sample, Disease_type, Cohort), by = "Sample") |> 
  mutate(Disease_type = factor(Disease_type, levels = c("Other", "No diagnosis", "Infectious", "Autoimmune", "Inflammatory", "Cancer")),
         Cancer = case_when(
           Disease_type == "Other" ~ "No_0", 
           Disease_type == "No diagnosis" ~ "No_1", 
           Disease_type == "Infectious" ~ "No_2", 
           Disease_type == "Autoimmune" ~ "No_3", 
           Disease_type == "Inflammatory" ~ "No_4",
           Disease_type == "Cancer" ~ "Yes")) |> 
  mutate(Cohort = factor(Cohort, levels = c("MEDECA", "ALLVOS")),
         Assay = factor(Assay, levels = proteins)) |> 
  filter(Cohort == "MEDECA") |> 
  ggplot(aes(Cancer, NPX, fill = Disease_type, color = Disease_type)) +
  geom_quasirandom(size = 0.5, show.legend = F) +
  geom_boxplot(alpha = 0.5, outlier.color = NA, show.legend = F) +
  stat_summary(fun = "median",
               geom = "crossbar", 
               width = 0.2,
               colour = "grey20",
               show.legend = F) +
  facet_grid(Assay~Cohort, scales = "free") +
   scale_color_manual(values = c(pal_controls, "Cancer" = "#E16C54")) +
  scale_fill_manual(values = c(pal_controls, "Cancer" = "#E16C54")) +
  theme_hpa()

ggsave(savepath("proteins_medeca.pdf"), height = 8, width = 3)

data |> 
  filter(Assay %in% proteins,
         Sample %in% metadata$Sample) |> 
  left_join(metadata |> 
              select(Sample, Disease_type, Cohort), by = "Sample") |> 
  mutate(Disease_type = factor(Disease_type, levels = c("Other", "No diagnosis", "Infectious", "Autoimmune", "Inflammatory", "Cancer")),
         Cancer = case_when(
           Disease_type == "Other" ~ "No_0", 
           Disease_type == "No diagnosis" ~ "No_1", 
           Disease_type == "Infectious" ~ "No_2", 
           Disease_type == "Autoimmune" ~ "No_3", 
           Disease_type == "Inflammatory" ~ "No_4",
           Disease_type == "Cancer" ~ "Yes")) |> 
  mutate(Cohort = factor(Cohort, levels = c("MEDECA", "ALLVOS")),
         Assay = factor(Assay, levels = proteins)) |> 
  filter(Cohort == "ALLVOS") |> 
  ggplot(aes(Cancer, NPX, fill = Disease_type, color = Disease_type)) +
  geom_quasirandom(size = 0.5, show.legend = F) +
  geom_boxplot(alpha = 0.5, outlier.color = NA, show.legend = F) +
  stat_summary(fun = "median",
               geom = "crossbar", 
               width = 0.2,
               colour = "grey20",
               show.legend = F) +
  facet_grid(Assay~Cohort, scales = "free") +
   scale_color_manual(values = c(pal_controls, "Cancer" = "#E16C54")) +
  scale_fill_manual(values = c(pal_controls, "Cancer" = "#E16C54")) +
  theme_hpa()

ggsave(savepath("proteins_allvos.pdf"), height = 8, width = 3)
```



# Comparison DE / ML

```{r}
medeca_de |> 
  left_join(ml_medeca_filter$important_proteins , by = c("Assay" = "Variable")) |> 
  filter(!is.na(Importance)) |> 
  mutate(Type = case_when(Importance > 0 & Sign == "POS" & logFC > 0 ~ "Both up",
                          Importance < 0 & Sign == "NEG" & logFC < 0 ~ "Both down",
                          T ~ "Different sign")) |> 
  ggplot(aes(Importance, logFC, color = Type)) +
  geom_vline(xintercept = 0, lty = "dashed", color = "grey") +
  geom_hline(yintercept = 0, lty = "dashed", color = "grey") +
  geom_point() +
  geom_text_repel(aes(label = Assay), show.legend = F) +
  theme_hpa() +
  ggtitle("Protein importance VS fold change - MEDECA")
```
